requirements.txt
# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.3

# LLM
anthropic==0.28.0
httpx==0.24.1

# API Backend
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Frontend
streamlit==1.28.2

# Utilities
pydantic==2.5.0